streamlit_app/requirements.txt
torch==2.3.1
torchvision==0.18.1
streamlit==1.36.0
pandas
pillow
opencv-python-headless
matplotlib
